Cleanup CommonExtensions
